@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "found": "^0.5.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1"
+    "found": "^1.1.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@vitejs/plugin-react": "^2.0.1",
+    "vite": "^3.0.7"
   },
   "scripts": {
-    "start": "react-scripts start"
+    "start": "vite"
   },
   "browserslist": [
     ">0.2%",
